Adds transliteration to keyword match
