+++ js/stringProcessing/findKeywordInUrl.js
@@ -1,7 +1,9 @@
 /** @module stringProcessing/findKeywordInUrl */
 
-var keywordRegex = require( "../stringProcessing/stringToRegex.js" );
+var matchTextWithTransliteration = require( "./matchTextWithTransliteration.js" );
+
 /**
+ * Matches the keyword in the URL.
  *
  * @param {string} url The url to check for keyword
  * @param {string} keyword The keyword to check if it is in the URL
@@ -13,7 +15,7 @@ module.exports = function( url, keyword ) {
 
 	if ( formatUrl !== null ) {
 		formatUrl = formatUrl[ 0 ].replace( /<.*?>\s?/ig, "" );
-		if ( formatUrl.match( keywordRegex( keyword ) ) !== null ) {
+		if ( matchTextWithTransliteration( formatUrl, keyword ).length > 0 ) {
 			keywordFound = true;
 		}
 	}
